* change the timeEvolution and produceInitialPopulationVector member functions in reaction.cpp in order to integrate a new member function, produceEquilibriumVector * perform timeEvolution calculations using long double precision, necessitating changes of variable from double to long double in some other functions * decrease the number of time steps for which output data is printed in mesmer.test
SVN: r159@trunk

+++ src/AssociationReaction.cpp
@@ -140,7 +140,7 @@ namespace mesmer
 
         // Get Boltzmann distribution for detailed balance.
         const int MaximumGrain = getEnv().MaxGrn ;
-        vector<double> adductPopFrac ; // Population fraction of the adduct
+        vector<long double> adductPopFrac ; // Population fraction of the adduct
 
         const int pdtGrainZPE = m_pdt1->get_grnZpe();
         m_pdt1->normalizedGrainDistribution(adductPopFrac, MaximumGrain - pdtGrainZPE) ;
